requirements.txt
pymongo
pydantic
python-dotenv
typing
dataclasses
fastapi